extract TextViewHolder from Text model
