+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/model/TextViewHolder.java
@@ -1,0 +1,28 @@
+package org.cru.godtools.tract.model;
+
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.annotation.UiThread;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import org.cru.godtools.tract.R;
+import org.cru.godtools.tract.R2;
+
+import butterknife.BindView;
+
+@UiThread
+final class TextViewHolder extends BaseViewHolder<Text> {
+    @BindView(R2.id.content_text)
+    TextView mText;
+
+    TextViewHolder(@NonNull final ViewGroup parent, @Nullable final BaseViewHolder parentViewHolder) {
+        super(Text.class, parent, R.layout.tract_content_text, parentViewHolder);
+    }
+
+    @Override
+    void onBind() {
+        super.onBind();
+        TextViewUtils.bind(mModel, mText);
+    }
+}
